[refactor] 채팅목록에서 채팅할사람을 고르는것을 상위컴포넌트에서 넘겨줌으로써 랜더링 최적화
