+++ src/components/Direct/Aside/ChatList/ChatListItem/ChatListItem.tsx
@@ -1,21 +1,9 @@
 import styled from "styled-components";
 import useGapText from "Hooks/useGapText";
-import { useAppDispatch, useAppSelector } from "app/store/hooks";
-import { selectChatItem, selectView } from "app/store/ducks/direct/DirectSlice";
-import { deleteRoom, makeRoom } from "app/store/ducks/direct/DirectThunk";
-import axios from "axios";
 import React from "react";
 import Direct from "../../../../../pages/Direct";
 
-const BASE_URL =
-    "http://ec2-3-36-185-121.ap-northeast-2.compute.amazonaws.com:8080";
 
-const token = {
-    accessToken:
-        "eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiIxIiwiYXV0aCI6IlJPTEVfVVNFUiIsImV4cCI6MTY0MzgxMTU3MH0._bLGXXtPlrAWXf8FVwGTGGeJSWb5S45tzqzatQQkYuUkZ0DzDiZJgi7GTgMerDhxmyms-PFTlL8HwueKqmdejg",
-    refreshToken:
-        "eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiIxIiwiYXV0aCI6IlJPTEVfVVNFUiIsImV4cCI6MTY0MzgwMzIwNH0.pftOV8QO0D9gEhIyJMtdQ13u-eUHzDKR4qmLOITb44Y-YERm_OyInkovsCrw4YgnSVfNAlP52uC8Y1bfIpXgOA",
-};
 
 interface ChatListItemContainerType {
     unreadFlag: boolean;
@@ -80,38 +68,17 @@ const ChatListItemContainer = styled.a<ChatListItemContainerType>`
 
 interface ChatListItemProps extends Direct.ChatItem {
     isSelected: boolean;
+    chatListClickHandler: (chatRoomId:number,username:string) => void
 }
 
-const ChatListItem = ({ chatRoomId, invitees, lastMessage, unreadFlag, isSelected }: ChatListItemProps) => {
+const ChatListItem = ({ chatRoomId, invitees, lastMessage, unreadFlag, isSelected ,chatListClickHandler}: ChatListItemProps) => {
     const calculatedTime = useGapText("2021.11.22 21:00:00");
-    const dispatch = useAppDispatch();
-    const view = useAppSelector(state => state.direct.view);
-    const chatListClickHandler = async () => {
-        // 클릭한거 읽은거였으면 읽음 처리해주는 로직 필요
-        dispatch(selectChatItem(chatRoomId));
-        if (view === "requests" || view === "requestsChat") {
-            dispatch(selectView("requestsChat"));
-        } else {
-            dispatch(selectView("chat"));
 
-            // 채팅방 클릭시 채팅방 생성 이 경우에는 기존에 목록에 있는 채팅방을 클릭하므로 실제 생성되진 않고, 기존의 Room 이 return 된다.
-            await dispatch(makeRoom({ token: token.accessToken, username: invitees[0].username }));
-
-            // 채팅방 클릭시 채팅방조회(채팅방을 클릭하면 unseen count를 감소시키는 API) 호출
-            const config = {
-                headers: { Authorization: `Bearer ${token.accessToken}` },
-            };
-            try {
-                await axios.delete(`${BASE_URL}/chat/rooms/${chatRoomId}`, config);
-            } catch (error) {
-                console.log(error);
-            }
-
-        }
-    };
     return (
         <ChatListItemContainer unreadFlag={unreadFlag} isSelected={isSelected}
-                               onClick={chatListClickHandler}>
+                               onClick={()=>{
+                                   chatListClickHandler(chatRoomId,invitees[0].username)
+                               }}>
             <img src={invitees[0].imageUrl} alt="avatarImg" className="user-image" />
             <div className="right-section-container">
